[CodeCompletion] Move CompletionLookup to its own file
