+++ include/swift/IDE/PossibleParamInfo.h
@@ -1,0 +1,37 @@
+//===--- PossibleParamInfo.h ----------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef SWIFT_IDE_POSSIBLEPARAMINFO_H
+#define SWIFT_IDE_POSSIBLEPARAMINFO_H
+
+#include "swift/AST/Types.h"
+
+using namespace swift;
+
+struct PossibleParamInfo {
+  /// Expected parameter.
+  ///
+  /// 'nullptr' indicates that the code completion position is at out of
+  /// expected argument position. E.g.
+  ///   func foo(x: Int) {}
+  ///   foo(x: 1, <HERE>)
+  const AnyFunctionType::Param *Param;
+  bool IsRequired;
+
+  PossibleParamInfo(const AnyFunctionType::Param *Param, bool IsRequired)
+      : Param(Param), IsRequired(IsRequired) {
+    assert((Param || !IsRequired) &&
+           "nullptr with required flag is not allowed");
+  };
+};
+
+#endif // SWIFT_IDE_POSSIBLEPARAMINFO_H
